+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created building access widget [#932](https://github.com/rokwire/illinois-app/issues/932).
 ### Changed
 - Open MyMcKinley web app in WebPanel instead of in external browser [#938](https://github.com/rokwire/illinois-app/issues/938).
+### Fixed
+- Show buss pass panel for residents [#936](https://github.com/rokwire/illinois-app/issues/936).
 
 ## [3.0.44] - 2021-11-09
 ###Fixed